nodejs quickstart updates for docs
