+++ documentation-samples/quickstarts/publish-knowledge-base/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "qnamaker-quickstart-publish-knowledge-base",
+  "version": "1.0.0",
+  "description": "",
+  "main": "publish-knowledge-base.js",
+  "scripts": {
+    "start":"node publish-knowledge-base.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
