+++ documentation-samples/quickstarts/update-knowledge-base/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "qnamaker-quickstart-update-knowledge-base",
+  "version": "1.0.0",
+  "description": "",
+  "main": "update-knowledge-base.js",
+  "scripts": {
+    "start":"node update-knowledge-base.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
